Cleaned up the code a bit, renamed the System in the IPhysics folder to PhysicsSimulationSystem, fixed the constraint creation so duplicates aren't created.

+++ src/voxelObjects/voxel.h
@@ -16,7 +16,7 @@ struct Voxel
   glm::vec3 mPosition{0.0f};
   glm::vec3 mPositionRelativeToCenter{0.0f};
   glm::vec3 mRelativePositionDelta{0.0f};
-  std::vector<glm::vec3> mNeighbors;
+  std::vector<glm::uvec3> mNeighbors;
   std::vector<u32> mMeshVertices;
 
   Voxel() = default;
